45. Insert Interval
You are given an array of non-overlapping intervals intervals where intervals[i] = [start_i, end_i] represent the start and the end of the i-th interval, and interval is another interval that needs to be inserted. Insert interval into intervals such that the list is still sorted and does not have any overlapping intervals.  Merge the new interval and adjust existing intervals.